Fix to replace spaces and other unaccepted chars from CFn stack names
With this fix it replaces all unaccepted chars from the CloudFormation
(CFn) stack name with a dash.

For example, when the OU name would be 'IT Department', it would fail
setting up the base ADF resources as it would generate the following
stack name: 'adf-global-base-IT Department', which includes a space that
is not supported by CloudFormation.

With this fix, it will replace the space with a dash, the new name would
become: 'adf-global-base-IT-Department'. Any other unsupported character
is treated the same.