Fix for AWS Account Ids prefixed with zeroes
**Issue:**

When the deployment map would refer to a specific AWS Account Id that
has a zero as the first digit, the number would be interpreted
differently or would not be accepted by validation at all.

For example, `012345678901` would become `12345678901` (zero removed),
and `012345671234` would become `1402434204` (interpreted as octal).

**Fix:**

In this commit, the schema validation for AWS Account Ids allows the use
of strings to define the account id. Such that an account id that has a
zero at the start would not be changed.

Additionally, the logic has been changed to detect AWS Account Ids that
do not match the AWS Account Id regular expression. In case this
happens, the error message will instruct the user to define the account
id in a string. It will also show how the account id was interpreted by
ADF and what the possible non-octal version of that number could be,
such that the user is able to find the faulty definition quickly.

As a side benefit, the default provider types in the schema validation
would return an unrelated error message when one of the properties did
not match. All of a sudden it would report that `s3` does not match
`codecommit`, for example. This has been fixed by matching the right
properties schema to the defined properties, based on the supplied
provider value.