Added mono repo support (#160)
Adds the `root_dir` property to the `cloudformation` deployment
provider properties. This enables the user to define multiple pipelines
that target the same repository, where each application/pipeline could
use different directories to host the CloudFormation template and
parameters.

In terms of CodeBuild: make sure to `cd` into the correct directory as
shown in the mono-repo sample. This is required to generate the
parameter files in the correct directory as well.

No changes were required for CodeBuild mono-repo support, as you can
specify the location of the BuildSpec using the `spec_filename` property
of the `codebuild` provider properties.

**R0915:**

A note on the added `pylint: disable R0915`: intentionally the modifications
required to add mono repo support touch as little as possible.
However, by adding a four statements we crossed the limit of
max 30 statements per function. The code in this file needs to be
refactored.

I will refactor the code in this specific file in a separate PR later.