Moved CodeBuild deployment action outside of region loop
**Issue:**

In the current format, one could think that CodeBuild supports being
deployed in a specific target region while this is not supported.

As written in #168, the name of the CodeBuild project would clash.

When CodeBuild steps are added as deployment actions, they should be
added as one per stage, not multiple in the same stage.

**Fix:**

Moved the CodeBuild deployment action outside of the region loop to
ensure that it will only create just one even though multiple regions
are configured as targets for this pipeline.

This would allow targeting multiple deployment regions for other
deployment steps, while only using one CodeBuild action as a deployment
stage.

**Why not adding multi-region support?**

The pipeline is deployed using CloudFormation in the main region.
Adding resources that get deployed in other regions from the same stack
would make things too complex to maintain in the long run.