Use pipeline name for CodeBuild instead of target name
**Issue:**

Currently, the name of the CodeBuild project is determined based on the
name of the target it points too. In the case of CodeBuild this would
always be the name of the deployment account.

Having multiple pipelines that have a CodeBuild project with the same
stage index would clash though.

**Fix:**

Use the name of the pipeline for CodeBuild
instead of the target name, as it will always
operate from the deployment account.