Fixes CodeBuild unique project name for deployment stages (#170)
* fixed broken actions

* Moved CodeBuild deployment action outside of region loop

**Issue:**

In the current format, one could think that CodeBuild supports being
deployed in a specific target region while this is not supported.

As written in #168, the name of the CodeBuild project would clash.

When CodeBuild steps are added as deployment actions, they should be
added as one per stage, not multiple in the same stage.

**Fix:**

Moved the CodeBuild deployment action outside of the region loop to
ensure that it will only create just one even though multiple regions
are configured as targets for this pipeline.

This would allow targeting multiple deployment regions for other
deployment steps, while only using one CodeBuild action as a deployment
stage.

**Why not adding multi-region support?**

The pipeline is deployed using CloudFormation in the main region.
Adding resources that get deployed in other regions from the same stack
would make things too complex to maintain in the long run.

* Refactored stage name code for improved readability

* Use pipeline name for CodeBuild instead of target name

**Issue:**

Currently, the name of the CodeBuild project is determined based on the
name of the target it points too. In the case of CodeBuild this would
always be the name of the deployment account.

Having multiple pipelines that have a CodeBuild project with the same
stage index would clash though.

**Fix:**

Use the name of the pipeline for CodeBuild
instead of the target name, as it will always
operate from the deployment account.