Fix to skip deletion of deployment_maps dir and other json/yaml files. (#161)
**Issue:**

When upgrading ADF, it deletes the old files that have been removed
between versions. However, the way it performs this test contained an
issue where it deleted the `deployment_maps` directory and other json/yaml
files that were not specifically whitelisted to be kept.

**Fix:**

Checks the files that would be scheduled for deletion against a regular
expression. When the file ends with a `json`, `yaml`, `yml` file extension
it should be kept in the repo as is.

Unit tests have been added to ensure that it would work as intended.

**Duplication:**

The `initial_commit.py` file is present twice with a almost completely
same setup. I have ensured the two files are even more alike, i.e. code
layout/white space fixes. This makes it easier to compare the two.

However, we would need to get this deduplicated later on.