Adding in a pipeline type param.
Changing the logic of cdk_stacks/main.py so that it will use that param
in order to select the right type of pipeline to deploy.

Also added in some unit tests to ensure it throws an exception when an
invalid type is used as well as some unit tests to ensure that the
adf_standard_pipeline type works as intended. As it has now been moved
into its file so that main.py can be easily extended to add new pipeline
types.

Initial Commit at seperating out pipeline generation from main.py

Unit tests

Documentation Update