Fix IAM for retrieve organization accounts helper
The `adf-build/helpers/retrieve_organization_accounts.py` script will need
to determine which organization it belongs to.

It performs this operation using the `organizations:DescribeOrganization`
API call [0] from inside CodeBuild in the deployment account.

In order for this to be successful, the deployment account needs to
grant CodeBuild access to this operation. This change will add that
to the policy of CodeBuild.

**References:**

* [0](https://docs.aws.amazon.com/organizations/latest/APIReference/API_DescribeOrganization.html)