Fix import references of export to output key
In the code and docs of the import statement, it was referring to the
`export_key`. However, in the fine print it stated that it is referring
to the `output_key` as part of the Stack outputs instead.

This was confusing people to create outputs with export ids, thinking
they had to. This update renames the usage to `output_key` instead such
that it is consistent in the docs and code.