Fix setting CodeBuild image strings in targets
**Issue:**

When one would try to specify the CodeBuild image in the target
environment when using CodeBuild, it did not accept having strings
as input.

For example, with this deployment map:
```yaml
  - name: sample-pipeline
    default_providers:
      source:
        provider: codecommit
        properties:
          account_id: 111111111111
      build:
        provider: codebuild
        properties:
          image: STANDARD_3_0
      deploy:
        provider: cloudformation
        properties:
          action: replace_on_failure
    targets:
      - name: deploy-to-dev
        path: /development
      - name: integration-tests-on-dev
        provider: codebuild
        properties:
          image: STANDARD_3_0
          spec_filename: testspec.yml
```

This would fail with the following error:
```
jsii.errors.JSIIError: Expected object reference, got "STANDARD_3_0"

Subprocess exited with error 1
```

**Fix:**

Refactored the `determine_build_image` function to accept both and reduced
the complexity along the way. Additionally, tests have been written to cover
all the different use cases for this function.