Fix error handling of generate_params intrinsic upload function
**The issue:**

When uploading files using the intrinsic upload function, and referring
to a file that does not exist, the error would go unnoticed. Causing the
parameter files to include references to S3 objects that do not exist.
Which one would only realize upon deployment of the template/when using
the referenced object.

**What is fixed?**

The upload code in the `S3` class has been refactored to improve
readability and reusability. New test cases ensure that missing file issues
and other errors are handled correctly.