Use isort v4
With the release of isort v5 the builds of our main branch kept failing.
Travis CI was using the latest version as it was not instructed to do
otherwise.

With this release, we force it to install v4 instead so builds continue to
operate until they support isort v5 in pylint.