Fix #280 - create deployment account concurrency failure
**Why?**

When the account creation failed due to a concurrency failure, it would
retry creating the account. However, since the retry was not returned
the created account would be ignored and the original invocation would
fail with `UnboundLocalError: local variable 'create_account' referenced
before assignment`.

**What is changed?**

The logic to create the account is separated into separate functions to
support testing the code segments. Tests have been written to simulate
the situation that failed before. Since no tests were written for this
custom resource yet, the other code that was part of the ensure_account
function got tests to ensure their operation too.

This fixes #280.