Remove dependency on botocore.vendored.requests
**Why?**

Releases of botocore will no longer support the requests library as part
of its vendored libraries. The samples/sample-expunge-vpc code was
relying on this dependency indirectly.

**What?**

The samples/sample-expunge-vpc code included the crhelper library
directly. This change will remove the dependency from being committed
into this repository and install it on the fly instead.

The latest update of the crhelper is removing the dependency of the
botocore.vendored.requests library. So by changing this to
update/install on the fly it will resolve issue #324.