Lock down buckets created by ADF - no public access (#350)
**Why?**

By default, ADF does not mark any object or bucket to be accessibly
publicly. However, the buckets did not restrict any other process from
doing so anyway.

**What?**

To ensure objects are not accidentally made public, this change will
enforce that. Changing the object or bucket policies to public would not
be allowed.

Additionally, one bucket, the BootstrapArtifactStorageBucket that was
added recently by #270 did not configure encryption and versioning.
This was not released as part of ADF yet, so no data was exposed.

The samples that deployed buckets are also updated to include the
encryption, versioning, and block-public access properties.