Add ability to disable trigger on changes for source providers (#357)
**Why?**

For several use cases, it would be great to have a pipeline that
would not trigger on source updates.
For example, a pipeline might only need to be triggered on a
schedule, by the completion of another pipeline, or manually.

This was requested by the following issues:
* Issue #308 - Starting pipeline only on completion_trigger
* Issue #337 - Question: Disable webhook for Github provider

**What?**

The CodeCommit, Github, and S3 source providers have been update
to support disabling triggering on source changes.

As described in the `docs/providers-guide.md` documentation, you
can optionally disable it when required.

This code also solves a bug when a pipeline would rely on polling
of changes for CodeCommit. In the previous version, it would be
triggering twice instead of once when the source account id was
specified. As the event handler would be listening for events
that indicate an update was performed, even when the pipeline
was set to use polling for changes instead of events.