Upgrade dependencies (CDK, Pylint, and others) (#364)
* Fix duplicate notification endpoint setup in pipeline generation

**Why?**

When generating a pipeline, the code optimization introduced by PR #285
broke the notification_endpoint configuration. When set, it would try to
configure the notification in the main generator as well as in the
default ADF pipeline generator.

**What?**

Removing the notification endpoint configuration from the main template,
as that is replaced by the default ADF pipeline generator
implementation.

* Upgrade to latest CDK, pylint, and others

Upgrade to CDK v1.105 and latest version of pylint and other
libraries.

Upgrade of pylint resulted in new errors as listed below.
These errors have been fixed as part of this commit too.