Fix permission to set Support Subscription upon account creation
**Why?**

As reported in #379, the IAM permission required to enable Support
Subscriptions upon account creation is missing.

**What?**

This change adds another Role that is used by the Bootstrap CodeBuild
project to request the right support level upon account creation.

A separate role is required, as the permissions should not be added to
the default ADF CodeBuild role that is also used by the pipelines
generated by ADF.