Fix use of separate container image per target
**Why?**

As reported by #382, setting a specific container image to use by the
CodeBuild provider to deploy to a target resulted in an error.

The CDK stack would fail to synthesize as the 'custom_repo' resource
name was already present in the stack.

**What?**

This fix ensures a unique identifier is used for custom container repositories.