Set Log Level as part of the SAR deployment process
**Why?**

When you run into errors deploying the latest version of ADF, it was not
easy to enable the debug level, as those used to be hard coded and
non-configurable.

**What?**

This change introduces a Log Level parameter to the SAR deployment
process. Allowing the administrators to deploy ADF with a specific Log
Level such as DEBUG directly from the SAR.