Add retry logic to SF Lambda invocations
**Why?**

When the Lambda service is experiencing issue, it could affect the
execution of the Step Function. Most of the errors it would fail with
could be recovered from.

**What?**

Added retry logic around Lambda service issues. Plus added a retry
mechanism to recover from Lambda function executions that ran for too
long. To retry those automatically too.