Fix duplicated steps in Account Bootstrap Step Function (#414)
**Why?**

In one of our recent PRs the indentation of the state machine was
changed. However, with this update, it seems that part of the state
machine got copied.

Because of this, Step Function was unable to deploy/update the state
machine.

**What?**

Removed the redundant parts of the state machine.
I checked both definitions, the repeated segments are a 1:1 copy.