Change example email domains and account ids
**Why?**

To ensure that email addresses or account ids are not exposed by accident,
this is the first step to ensure that the same example email addresses
and account ids are used. Such that a mechanism can be introduced to
prevent accidentally committing email addresses or account ids that are
not supposed to be share publicly.

**What?**

This step updates all the email addresses and account ids in the code and
documentation.

The instructions for our contributors received an update too.

In case you want to specify an example email address, please make use of
one of the following email addresses:

* `jane@example.com`
* `john@example.com`
* `janes_team@example.com`
* Or another `@example.com`
* For tests that could potentially create accounts (if mocks failed to work),
  please use `test@amazon.com` instead.

When you want to write documentation/tests and need an example account id, the
following account ids may be used:

* `111111111111`
* `222222222222`
* ... to ...
* `999999999999`
* or to show the length more easily:
* `012345678910`
* `012345671234`
* `123456789012`