Move account process to main SAR template
**Why?**

The SAR does not understand local references as introduced by the
src/account_processing.yml unfortunately.

**What?**

After investigating several options, the only option that was accepted was to
move the logic inside the main template instead.
The references are updated to use the ones available in the main template
directly.

Additionally, the ARN references are updated to include the partition correctly.