Fix: Wave deployment input data
**Why?**

When generating the waves, it does so in the input generation.
When it was shared to the actual deployment of the CodePipeline stack,
the data types did not match.

**What?**

This pull request adds comments and restructures the wave generation at
the input side to make it work.

Co-authored by: StewartW