feat: enable use of codebuild reports
Currently, you cannot use the reports functionality in the build phases of your pipelines. This has to do with the fact that the IAM Role that is executing the build has no permissions to do so. This prevents you from collecting reports using the CodeBuild Reports functionality. By adding the needed IAM actions to the CodeBuild Role you will be able to collect your reports for unit tests and coverage reports.