feat: allow adf-global-base-deployment parameter overwrites
We use a prefix in the CodeCommit repo name. So all our stacks are already prefixed with that prefix.

For this reason we changed the StackPrefix to an empty string in the global.yml for the deployment account. This could not be done without an customization, so we removed the `adf-` prefix in the template.

This pull request adds support for setting parameters in a JSON file. I an not sure about the format of the file itself. Or that this will work in this form. Lets use this pull request to figure out what the best way is.