feat: allow `TemplateConfiguration` override from the target properties
By specifying a `configuration_file_path` property on the target properties you control the configuration used. This could be useful when you do not want to use the `generate_params.py` script. Or if you have multiple stacks with different parameters that you want to deploy from a single pipeline.

Issue: #449