Chore: Add editorconfig to repo
**Why?**

With editorconfig, editors that are used across our collaborators will
be able to use the same indentation / line lengths and other configuration
settings.

You can find more information on how to install the editorconfig add-on for your
editor on: https://editorconfig.org/

**What?**

Adding specification as currently applied in the files in our repository.