Fix: Lookup correct Organization API Region
**Why?**

At the moment, the organization API calls are made to the region where
the organizations.py code is running.

However, which region to use depends on the partition that is in use.
For the gov-cloud regions, it should use the `us-gov-west-1` region.
While the `aws` partition regions should use `us-east-1`.

**What?**

The code will determine the partition and return the correct region to
use for the Organizations API.