Use boto3 Session to determine partition + throw error if unsupported
**Why?**

Use of the boto3 Session get_partition_for_region function allows us to
rely on a supported and maintained function instead.

Certain partitions are not supported by ADF yet. If such a partition is
detected, we should throw an error to highlight this.

**What?**
Refactored to boto3 get_partition_for_region for partition lookup.

Throwing error if the partition is not supported by ADF yet.