Fix update process to only flag helpers as executable (#499)
**Why?**

Multiple files were marked as executable without a specific reason to be
executable. A permission change in our repository easily goes unnoticed when
a pull request is opened.

**What?**

With this change, we keep track of the files that should be marked executable
and only flag those as such. If a file would end up in the ADF repo with the
wrong permissions, this will not get replicated to the bootstrap / pipelines
repository as managed in ADF user environments.

This change also updates the permissions of files that were marked executable
without a good reason to be executable.