Add missing permission to assume Cross-Account-Org R/O role
**Why?**

The Generate Pipeline Input lambda function tried to assume into the
cross-account access role, to read the organization API.
This was not allowed, as it was not permitted by the cross-account access role
as a trusted role that can assume into the read-only role in the management
account.

**What?**

Addressed by adding a specific role name for the role that is used by the
Generate Input lambda function. Plus putting that role on the allowed list of
roles that are allowed to assume the cross-account access role for read-only
access to the organization API.