Add reference to state machine from CodeBuild on pipeline updates
**Why?**

In the past, the ADF Pipelines were updated from the CodeBuild step in the
aws-deployment-framework-pipelines CodePipeline flow.

With the recent release of the Step Function Pipeline Management State Machine,
this logic has moved.
To ensure that administrators still understand what happens and whether the
pipelines are updated correctly, we should link them to the state machine from
the CodeBuild execution.

**What?**

Doing exactly that, adding a reference to the state machine in the CodeBuild
step.