Improvement: Only invoke pipeline deletion when needed (#510)
**Why?**

At the moment, the InvokeDeleteStateMachine task is called even when no
pipelines need to be deleted. This causes the actual state machine that deletes
these tasks to fail. As it cannot iterate over an empty list.

**What?**

Introducing a choice, that will only invoke the delete pipelines state machine
when there is a need to do so.