Add support to change main branch on bootstrap and pipelines repositories (#508)
* Add support to change main branch on bootstrap repo

**Why?**

We set a goal to remove the word `master` from our repository as much as
possible. The bootstrap repository initialized to the `master` branch by
default and creates pull requests against the `master` branch too.

Additionally, the CodePipeline should monitor the correct branch to
perform the changes required when the main branch gets updated.

**What?**

For any new installation, the bootstrap repository should default to the
`main` branch.

While any existing installation of ADF should allow the administrators
to move to the `main` branch easily.
To enable this, when ADF is updated, it will perform a check which branch is
configured as the default branch on the repository (adf-bootstrap and
adf-pipelines). It will create the pull request against the default branch if it
needs to be updated.

If you want to switch from one branch to another,
you need to create a new branch from the current default branch.

For example, switching from `master` to `main`:
Simply running these commands from inside the
`aws-deployment-framework-bootstrap` repository that is cloned locally
should do it:

```bash

git fetch
git checkout origin/master
git push HEAD:origin/main
```

Navigate to the CodeCommit repository and update the default branch of the
repository to the new branch. Make sure to click the `Save` button underneath
the default branch setting to save it.

* Add support for default branches on ADF pipelines repo

Apply default branch determination logic to the ADF-pipelines repository too.