Add retry logic to SF Lambda invocations (#513)
* Improve retry mechanism in account bootstrap process

**Why?**

The Lambda function in the account bootstrap times out when
on-boarding a large number of regions at once.

In some cases, this happens just after the CloudFormation Change Set
is created, leaving the stack in `REVIEW_IN_PROGRESS` state.

The next time the Step Function for account bootstrapping would
run, it detects that the stack exists and attempts to update it.
While in fact, it still needs to be created.

**What?**

This set of changes will improve the log messages, to ease
troubleshooting these issues in the future.

It will detect if the stack is in a clean state, if so it will update.
If not, it will remove the stack and recreate it instead.

The same holds for any Change Sets that were left on the Stack from a
prior execution. These will get deleted before a new one is created, as
these might contain outdated changes.

To ensure this error will not happen frequently, the Lambda function
will throw an error when there is not enough time to handle another
deployment. Such that the Step Function can retry with sufficient time
available to perform updates at the next attempt.

* Raise an error when the execution time is not sufficient

* Add retry logic to SF Lambda invocations

**Why?**

When the Lambda service is experiencing issue, it could affect the
execution of the Step Function. Most of the errors it would fail with
could be recovered from.

**What?**

Added retry logic around Lambda service issues. Plus added a retry
mechanism to recover from Lambda function executions that ran for too
long. To retry those automatically too.

* Fix indentation of State Machine definitions