Fix updating account alias when needed
**Why?**

At the moment, the account management state machine will attempt to add the
specified alias to the account whether it is present or not.

In case the alias already exists, it will fail.

**What?**

This change will ensure the aliases configured on the account are removed when
they are no longer configured (you can only have one, but the API supports
multiple). It will then add the alias if the alias that needs to be present is
missing.