Add support for the CodeBuild provider to run inside a VPC (#517)
**Why?**

To add support for build/deploy steps that use CodeBuild and need to
connect to internal resources that are only available through a VPC.

This feature allows you to configure which VPC to connect to, using a
specific list of subnets and security groups. The latter helps ensure
that the connection is only allowed to connect when a specific security
group is used.

**What?**

Added three properties to the CodeBuild provider:
- vpc_id (string)
- subnet_ids (list of ids (strings))
- security_group_ids (list of ids (strings))

When you configure a pipeline to use a specific VPC, you need to specify
the subnets to use as well.

In case you don't specify any security groups, they will be
automatically created for you. These security groups allow all egress
traffic by default. It is recommended to use a specific security group
so you can strictly define what egress traffic is allowed.

Please note: When you configure a specific CodeBuild environment in a
pipeline to use a VPC it cannot be removed easily. To remove the VPC
from the CodeBuild environment, you need to remove the pipeline stack
and instruct ADF to generate the pipelines again (release trigger on
aws-deployment-framework-pipelines in CodePipeline).

Documentation has been added and a sample that shows how to setup a
pipeline to use a specific VPC.