Reverting Pipeline Generation restrictions
**Why?**

With a recent PR, an attempt was made to restrict the `iam:PassRole`
permissions for the pipeline generator. However, since ADF supports configuring
which role to use, this would introduce a breaking change.

**What?**

Reverting to the previous policy of `iam:PassRole`.

Plan is to address this in the next major ADF release.