Forward CodePipeline Execution Id to Account Mgmt Creation SFN
Step 2 of fixing #518.

**Why?**

As explained in #518, we need to forward the execution id of the CodePipeline
that triggered the Account Management state machine so we can wait for to
complete.

**What?**

Adding the CodePipeline execution identifier to the Step Functions
State Machine invocation to enable tracing state machines in progress at
CodeBuild execution time.