Add ADF Version to S3 file sync, such that an ADF update triggers updates
**Why?**

When files are synced to S3, they only triggered an update of the account
management or pipeline generator when the file content changed.

If ADF made changes to the pipeline structure, the pipelines and account
management should be retriggered to apply them.

**What?**

By adding the `adf_version` metadata to the files that are synced, we can
ensure that we only trigger an update to the file when the version is updated.