Only process account and deployment maps generated by same ADF version
**Why?**

When an account file or deployment map is updated, it should only be processed
when the version equals the current ADF version. Otherwise it should skip the
file.

This will ensure we don't get into compatibility issues, where a file structure
update will make the processing of the file fail.

**What?**

Checking the version number attached to the S3 object metadata against the
current ADF version number. If those mismatch, it will skip the file.